+++ packages/agent_dart_base/test/identity/p256.dart
@@ -1,10 +1,7 @@
 import 'package:agent_dart_base/agent_dart_base.dart';
 import 'package:test/test.dart';
-import '../test_utils.dart';
 
 void main() {
-  ffiInit();
-  // matchFFI();
   p256Test();
 }